updated bot architecture
changed bot architecture
