+++ src/events/index.ts
@@ -18,6 +18,8 @@ export const botEvents: EventReturnType[] = [
     on, once
 ]
 
+export default botEvents;
+
 // for(const botEvent of botEvents){
 //     const fcn = client[botEvent.name];
 //     fcn(...botEvent.execute[0]);
