+++ src/events/on/index.ts
@@ -2,9 +2,11 @@ import { ClientEvents } from "discord.js";
 import { EventArguments, EventReturnType} from "../";
 
 import interactionCreate from "./interactionCreate";
+import messageCreate from "./messageCreate";
 
 const executeList = [
-    interactionCreate
+    interactionCreate,
+    messageCreate
 ] as EventArguments<keyof ClientEvents>[];
 
 const once: EventReturnType = {
