+++ src/events/on/messageCreate.ts
@@ -1,25 +1,125 @@
-const debug = require("debug"); debug("Event:MessageCreate");
+const debug = require("debug"); debug("ChrezBot:MessageCreate");
 
-import {prefixes} from "@config";
+import {max_message_allowed, muted, ownerID, prefixes} from "@config";
+import * as sharedCommands from "shared/commands";
 
 import {EventArguments} from "../"
+import { CommandBuilder } from "@modules/CommandBuilder";
+import { userIsAdmin } from "@modules/profiles";
+import { Cause } from "@typings/customTypes";
+import { sendError } from "@bot";
 
-const event: EventArguments<"messageCreate"> = ["messageCreate", (message) => {
-           // Chrez time (need spacing)
-           let prefixFlag = false;
-           for(const prefix of prefixes)
-               if(message.content.split(" ")[0].startsWith(prefix)) prefixFlag = true;
-           if(!prefixFlag) return;
-   
-           // remove the prefix and make each word as parameter
-           let params = message.content.trim().split(/ +/);
-           params.shift();
-   
-           // first argument is the command                       v assert, it have atleast 1 array so it's safe
-           const command = params.length == 0 ? "": params.shift()?.toLowerCase();
-   
-           console.log(message.content);
-           console.log(command);
+const event: EventArguments<"messageCreate"> = ["messageCreate", async (message) => {
+    if(!message.guild) return;
+
+    // check if the bot can send message into the channel
+    if(!message.guild.members.me?.permissions.has("ManageMessages")) return;
+    
+    // ignore message from bot or long message
+    if(message.author.bot || message.content.length > max_message_allowed) return;
+    
+    // inline command handling
+    if(!muted)
+    for(const [v, k] of sharedCommands.aliasCriteriaMap){
+        if(typeof v === "string")
+        if(message.content === v){
+            sharedCommands.inlineCommands.get(k)?.execute(message);
+            return;
+        }
+        if(v instanceof RegExp)
+        if(message.content.match(v) !== null){
+            sharedCommands.inlineCommands.get(k)?.execute(message);
+            return;
+        }
+    }
+    
+    // check if command is directed to chrezbot (e.g "Chrez" math)
+    if(prefixes.find(prefix => message.content.startsWith(prefix)) === undefined) 
+    return;
+    
+    /**
+     * argument variables, guaranteed lowercase and command removed
+     */
+    const args: string[] = message.content.split(/ +/);
+    args.shift(); // remove command name (Chrez)
+    
+    debug(`got message: ${message.content}`);
+    debug(`args: ${args}`);
+    
+    // check if command available (i.e: not just Chrez tho it should be alr guarded with inline command)
+    const command = args.shift();
+    // change args to lowercase
+    args.forEach((str, index, arr) => arr[index] = str.toLowerCase());
+    if(command === undefined) return;
+
+    try{
+        // check command
+        if(sharedCommands.command.has(command))
+            await sharedCommands.command.get(command)?.execute(message, args);
+        // check alias for command
+        else if(sharedCommands.commandAlias.has(command))
+            sharedCommands.command.get(sharedCommands.commandAlias.get(command)!)?.execute(message, args);
+        // check if command is for private members (highest authority)
+        else if(sharedCommands.privateCommands.has(command)){
+            if(message.author.id === ownerID || userIsAdmin(message.author.id))
+                await sharedCommands.privateCommands.get(command)?.execute(message, args);
+            else
+                throw new Error("This command is for private members only!");
+        }
+        // check if command is for private members (lower authority)
+        else if(sharedCommands.privateCommandAlias.has(command)){
+            if(message.author.id === ownerID || userIsAdmin(message.author.id))
+                await sharedCommands.privateCommands.get(sharedCommands.privateCommandAlias.get(command)!)?.execute(message, args);
+            else
+                throw new Error("This command is for private members only!");
+        }
+        else if(sharedCommands.allCommands.has(command)){
+            const chatCommand = sharedCommands.allCommands.get(command)!;
+
+            if(CommandBuilder.isCommandBuilder(chatCommand)){
+                if(chatCommand.status === "private"){
+                    const userId = message.author.id;
+                    if(userId === undefined){
+                        await sendError(message, "Cannot verify the sender of this command!");
+                        return;
+                    }
+                    else if(!userIsAdmin(userId)){
+                        await sendError(message, "This command is only available for private members!");
+                        return;
+                    }
+                }
+
+                const res = await chatCommand.execute(message, args);
+                if(Cause.isCause(res) && !res.ok)
+                    sendError(message, res.message);
+                        
+            }
+        }
+        else
+            throw new Error("Your command is not available in Chrez Command List");
+    }
+    catch(e: unknown){
+        try{
+            debug(`error: ${e}`);
+
+            // check if error can be send through discord
+            if(!message.channel) return;
+
+            if(e != null && typeof e === "object")
+                if("message" in e && typeof e.message === "string")
+                    sendError(message, e.message);
+                else if(typeof e == "string")
+                    sendError(message, e);
+                else
+                    sendError(message, "Unknown error occured");
+            else
+                sendError(message, "Unknown error occured");
+        }
+        catch(e: unknown){
+            // this catch is last resport if fatal error occured
+            console.log(`fatal error: ${JSON.stringify(e)}`);
+        }
+    }
 }]
 
 export default event;
